project upload - planet generators
brought back my planet generators in their state approx of when i had built them for my sccsunity vr portefolio but without VR this time around. for the pickaxe tip to shoot a projectile in my planetgeneratorrev3, you have to click with the left mouse button on the game window and a per frame projectile is going to be sent towards the chunk.