update
i developped a bigger planet generation using chunks of size 64x64y64z. I also just developed replacing my integer buffers of firstvertlocationx/y/z to be float4 instead of integer so now instead of using 6 compute buffers, i use only 2.
fix on the foot collision with the floor during mouse click movement where the character was going straight in a line at any height of the voxels on the planet surface as if it was flying towards the clicked location, this is fixed.