more simple ik
I used what i learned from using Unity3d and having crashes that always break our scenes inspector variables where everything is missing and we have to rebuild everything from the grounds up over and over again. I got tired and i remembered that the user BitCrushed on the unity community forums had answered me with a project that always passes the test of the porting issues, where when objects are created from script they don't need to be setuped manually in the inspector, as seen here https://forum.unity.com/threads/need-help-on-sword-attack-concept.455461/#post-2984933

so this new version of my third person controller has inverse kinematics but the camera isn't setup right now at the right location. A small chunk is spawned using a compute buffer for the intmap and the first vertex of each face and dimensions of each faces and the cpu works to create the final vertices lists and triangles lists to create the mesh. If you want to change the terrain or type of objects that the ground sticks to, disable the planet generator in the inspector and create a big cube as the scene floor and tag it "terrain" and the sccsik will find the object that it needs the feet to be attached to, then move the character towards the object and the feet and the rotation of the ik character should match the object pivot point.