final/final chunk neighboor fix
ok lol, sorry about the many fixes for the neighboor check of chunks to not build faces inside out of the mesh, fixes that never worked. i think this is the final final fix though. it seems fine again on my left screen and the logic is sound. we just aren't breaking bytes yet.