update - foot collision 6 gravity fix
i developed a fix for the foot collision with the cubic planet for using 6 different gravity axis. it seems to be working much better now, although i would need to implement movement restrictions next in order for the character not to go through walls or go over holes in the voxels dependant on the gravity sides.