update - foot collision no physics
i developed a way to have the ik character rig rotating correctly to the planetgravity rotations, with using euler angles instead of setlookrotation that didn't seem like an instant rotation when using it or seemed like it was using a smoothdamp. The results are better now with the angles but only when crossing over another face with using the move forward button. moving backwards when changing sides doesn't work yet.